# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Proxmox Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a specific tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/{id}/summary:
    get:
      operationId: getTenantSummary
      tags: [tenants]
      summary: Get tenant VM summary (counts by status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # VMs
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: string
        userCount:
          type: integer
        vmCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - userCount
        - vmCount
        - createdAt
        - updatedAt
    CreateTenantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
      required:
        - name
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
    TenantSummary:
      type: object
      properties:
        tenantId:
          type: integer
        totalVms:
          type: integer
        runningVms:
          type: integer
        stoppedVms:
          type: integer
        pausedVms:
          type: integer
        totalUsers:
          type: integer
      required:
        - tenantId
        - totalVms
        - runningVms
        - stoppedVms
        - pausedVms
        - totalUsers
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        role:
          type: string
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        status:
          type: string
        vmCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - username
        - email
        - role
        - status
        - vmCount
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        role:
          type: string
        tenantId:
          type: ["integer", "null"]
        password:
          type: string
      required:
        - username
        - email
        - role
        - password
    UpdateUserBody:
      type: object
      properties:
        username:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        tenantId:
          type: ["integer", "null"]
        status:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
...