# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Proxmox Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  status: string;
  userCount: number;
  vmCount: number;
  createdAt: string;
  updatedAt: string;
}