# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Proxmox Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  fullName?: string | null;
  role: string;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  tenantName?: string | null;
  status: string;
  vmCount: number;
  createdAt: string;
  updatedAt: string;
}